Clean up editor window code and added confirmation dialog to dangerous operation

+++ Editor/Core/Components/Main/ShapeShifterEditorWindow.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Reflection;
+using Miniclip.ShapeShifter.Saver;
 using Miniclip.ShapeShifter.Skinner;
 using Miniclip.ShapeShifter.Switcher;
 using Miniclip.ShapeShifter.Utils;
@@ -14,6 +15,31 @@ namespace Miniclip.ShapeShifter
     class ShapeShifterEditorWindow : EditorWindow
     {
         private bool showConfiguration;
+
+        [MenuItem("Window/Shape Shifter/Open ShapeShifter Window", false, 'G')]
+        public static void OpenShapeShifter()
+        {
+            ShowNextToInspector(true);
+        }
+
+        private static void ShowNextToInspector(bool focus = false)
+        {
+            Assembly editorAssembly = typeof(Editor).Assembly;
+            Type inspectorWindowType = editorAssembly.GetType("UnityEditor.InspectorWindow");
+
+            GetWindow<ShapeShifterEditorWindow>(
+                "Shape Shifter",
+                focus,
+                inspectorWindowType
+            );
+        }
+
+        private void OnSelectionChange()
+        {
+            ShapeShifter.DirtyAssets.Clear();
+            ShapeShifter.CachedPreviewPerAssetDict.Clear();
+            AssetWatcher.ClearAllWatchedPaths();
+        }
 
         private void OnGUI()
         {
@@ -35,58 +61,55 @@ namespace Miniclip.ShapeShifter
 
             using (new GUILayout.VerticalScope())
             {
-                showConfiguration = EditorGUILayout.Foldout(showConfiguration, "Configuration");
-
-                if (showConfiguration
-                    && ShapeShifterConfiguration.Instance.DefaultConfigurationEditor != null
-                    && ShapeShifterConfiguration.Instance.ExternalConfigurationEditor != null)
-                {
-                    ShapeShifterConfiguration.Instance.DefaultConfigurationEditor.OnInspectorGUI();
-                    ShapeShifterConfiguration.Instance.ExternalConfigurationEditor.OnInspectorGUI();
-                }
-
-                AssetSwitcherGUI.OnGUI();
-                AssetSkinnerGUI.OnGUI();
-                ExternalAssetSkinnerGUI.OnGUI();
-                PreMergeCheckGUI.OnGUI();
+                OnShowConfigurationGUI();
+                OnShowComponentsGUI();
 
                 GUILayout.FlexibleSpace();
-                
-                if (GUILayout.Button("Restore missing assets"))
-                {
-                    AssetSwitcher.RestoreMissingAssets();
-                }
 
+                OnShowUtilOperationsGUI();
                 OnDangerousOperationsGUI();
             }
 
             Repaint();
         }
 
-        private void OnSelectionChange()
+        private void OnShowConfigurationGUI()
         {
-            ShapeShifter.DirtyAssets.Clear();
-            ShapeShifter.CachedPreviewPerAssetDict.Clear();
-            AssetWatcher.ClearAllWatchedPaths();
+            showConfiguration = EditorGUILayout.Foldout(showConfiguration, "Configuration");
+
+            if (showConfiguration
+                && ShapeShifterConfiguration.Instance.DefaultConfigurationEditor != null
+                && ShapeShifterConfiguration.Instance.ExternalConfigurationEditor != null)
+            {
+                ShapeShifterConfiguration.Instance.DefaultConfigurationEditor.OnInspectorGUI();
+                ShapeShifterConfiguration.Instance.ExternalConfigurationEditor.OnInspectorGUI();
+            }
         }
 
-        [MenuItem("Window/Shape Shifter/Open ShapeShifter Window", false, 'G')]
-        public static void OpenShapeShifter()
+        private static void OnShowComponentsGUI()
         {
-            ShowNextToInspector(true);
+            AssetSwitcherGUI.OnGUI();
+            AssetSkinnerGUI.OnGUI();
+            ExternalAssetSkinnerGUI.OnGUI();
         }
 
-        private static void ShowNextToInspector(bool focus = false)
+        private static void OnShowUtilOperationsGUI()
         {
-            Assembly editorAssembly = typeof(Editor).Assembly;
-            Type inspectorWindowType = editorAssembly.GetType("UnityEditor.InspectorWindow");
+            if (GUILayout.Button($"Force Save To {ShapeShifter.ActiveGameName}", StyleUtils.ButtonStyle))
+            {
+                AssetSaver.SaveToActiveGameSkin();
+            }
 
-            GetWindow<ShapeShifterEditorWindow>(
-                "Shape Shifter",
-                focus,
-                inspectorWindowType
-            );
+            GUILayout.Space(30);
+
+            if (GUILayout.Button("Restore Missing Assets"))
+            {
+                AssetSwitcher.RestoreMissingAssets();
+            }
+
+            PreMergeCheckGUI.OnGUI();
         }
+
 
         private static void OnDangerousOperationsGUI()
         {
@@ -103,18 +126,32 @@ namespace Miniclip.ShapeShifter
         {
             Color backgroundColor = GUI.backgroundColor;
             GUI.backgroundColor = Color.red;
-            if (GUILayout.Button("Remove all skins"))
+            if (GUILayout.Button("Remove all skins") && RemoveAllSkinsDisplayDialog())
             {
+                EditorUtility.DisplayProgressBar("ShapeShifter", "Deleting internal skins", 0.0f);
                 AssetSkinner.RemoveAllInternalSkins();
+                EditorUtility.DisplayProgressBar("ShapeShifter", "Deleting external skins", 0.3f);
                 ExternalAssetSkinner.RemoveAllExternalSkins();
+                EditorUtility.DisplayProgressBar("ShapeShifter", "Cleaning up git ignore", 0.6f);
                 GitIgnore.ClearShapeShifterEntries();
+                EditorUtility.DisplayProgressBar("ShapeShifter", "Deleting main skins folder", 1f);
                 if (Directory.Exists(ShapeShifter.SkinsFolder.FullName))
                 {
                     ShapeShifter.SkinsFolder.Delete(true);
                 }
+                EditorUtility.ClearProgressBar();
+                GUIUtility.ExitGUI();
             }
 
             GUI.backgroundColor = backgroundColor;
+        }
+
+        private static bool RemoveAllSkinsDisplayDialog()
+        {
+            return EditorUtility.DisplayDialog("ShapeShifter",
+                $"You are about to remove shapeshifter's skin folders.\n Your project assets will remain " +
+                $"the same as the current game skin ({ShapeShifter.ActiveGameName}).\n You will loose the other game skins",
+                "Continue", "Cancel");
         }
     }
 }
